enable testing by others
