+++ app/src/ingestion/la_policy/scrape/requirements.txt
@@ -1,0 +1,2 @@
+install-playwright
+playwright
